backend/planmytrip_backend/requirements.txt
asgiref==3.7.2
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Django==4.2.6
django-cors-headers==4.3.0
django-enumchoicefield==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.6
